requirements.txt
django
djangorestframework
pygments
psycopg2-binary
coreapi
pyyaml
